Merged run-browserify and run-minify. Added static-analysis command.
Merged the run-browserify and run-minify commands into a single command
since these will almost always be used together. When browserify is run,
the libraries are bundled up and the minified version + source map is
also generated.

Task names have been changed to remove the 'run-' part because this
wasn't really helpful for distinguishing anything. Maybe we can bring
this back if it's useful/needed down the road.

Also added a single task to run jshint.