Seperated browserify from minification.
Went ahead and seperated the browserify build step from the minification
and sourcemap generation step. This was done to make it remedy some
issues where uglify was expected sourcemaps for the GCL parser.

It's happy now, and we can be a bit more flexible with sourcemap
generation and minification this way.