Added gulp-react to the project and fixed some gulp tasks.
Some gulp tasks weren't reacting very well to react and it's jsx
templating setup. The issue was that jsx takes its own liberties with
the javascript language, liberties that jshint and istanbul aren't aware
of.

Enter gulp-react. gulp-react will basically do an in-stream
transpilation of jsx templates into regular javascript in such a way
that gulp can work with it. This solves our problem of jshint and
istanbul being confused by the new syntax.